Feat: arrange plants info boxes into classic 4x4
Note that, though my solution and TOP's are very similar, TOP wins
by a slim but appreciable margin: instead of giving each plant boxe
it's own class name, a single one is shared among the four, whereupon
a max-width can be given. Practically, this does the same thing as
targeting all the direct descendants of the parent container, but
in a more elegant and simple manner.